pyproject.toml
[tool.poetry]
name = "knot exploration"
version = "0.1.0"
description = "tools for representing, loading, and manipulating knots"
authors = [
  "Aaron Aster <aarongoodm@gmail.com>"
]
repository = "https://github.com/ajgoodm/knots"

[tool.poetry.dependencies]
python = "^3.10"
attrs = "^23.1.0"
imageio = "^2.31.5"

[tool.poetry.group.dev.dependencies]
pre-commit = "~3"
pytest = "~7"
pylint = "2.17.7"
mypy = "^1.6.1"
jupyter = "^1.0.0"
matplotlib = "^3.8.1"

[tool.mypy]
mypy_path = "./knots"
exclude = [
  '.*/test_.*\.py$'
]
